Fixed all build errors except 2
Fixed all errors pertaining to more unavailable usings and a constructor not overriding a no longer abstract class
