+++ CrusaderKingsStoryGen/Managers/LanguageManager.cs
@@ -4,6 +4,7 @@
 
 namespace CrusaderKingsStoryGen.Managers
 {
+    using CrusaderKingsStoryGen.Helpers;
     using System;
     using System.Collections.Generic;
     using System.IO;
