+++ CrusaderKingsStoryGen/Managers/SocietyManager.cs
@@ -4,6 +4,7 @@
 
 namespace CrusaderKingsStoryGen.Managers
 {
+    using CrusaderKingsStoryGen.Helpers;
     using CrusaderKingsStoryGen.Parsers;
     using CrusaderKingsStoryGen.ScriptHelpers;
     using System;
