+++ CrusaderKingsStoryGen/PropertyPageProxies/CultureProxy.cs
@@ -9,6 +9,8 @@ namespace CrusaderKingsStoryGen.PropertyPageProxies
     using System.ComponentModel;
     using System.Drawing;
     using CrusaderKingsStoryGen.Forms;
+    using CrusaderKingsStoryGen.Helpers;
+    using CrusaderKingsStoryGen.Managers;
     using CrusaderKingsStoryGen.Parsers;
 
     class CultureProxy
