+++ CrusaderKingsStoryGen/PropertyPageProxies/ReligionProxy.cs
@@ -7,6 +7,7 @@ namespace CrusaderKingsStoryGen.PropertyPageProxies
     using System.ComponentModel;
     using System.Drawing;
     using CrusaderKingsStoryGen.Forms;
+    using CrusaderKingsStoryGen.Managers;
     using CrusaderKingsStoryGen.Parsers;
 
     class ReligionProxy
