+++ CrusaderKingsStoryGen/ScriptHelpers/ScriptScope.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using CrusaderKingsStoryGen.Helpers;
+using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
